making some changes to input handler etc.

+++ projects/Threading/Semaphore.h
@@ -12,13 +12,13 @@ namespace Threading
     class Semaphore
     {
     public:
-        Semaphore(uint32_t initial = 0, uint32_t max = std::numeric_limits<uint32_t>::max());
+        Semaphore(uint32_t initial = 0, size_t max = std::numeric_limits<size_t>::max());
         virtual ~Semaphore();
 
         bool init();
         bool wait();
         bool timedWait(uint32_t timeout = -1 /* milliseconds */);
-        bool post(uint32_t count = 1 /* how many times to post */);
+        bool post(size_t count = 1 /* how many times to post */);
         bool destroy();
 
     private:
